lazy load wdl dependencies
