+++ cirro/models/process/__init__.py
@@ -1,0 +1,1 @@
+from cirro.models.process.process import PipelineDefinition, get_input_params
