+++ cirro/models/process/process.py
@@ -6,7 +6,6 @@ from typing import Any, Iterable, Optional
 import logging
 
 from referencing import Resource, Registry
-import WDL
 
 
 class PipelineDefinition:
@@ -57,6 +56,10 @@ class PipelineDefinition:
                 else:
                     # otherwise, just take the first WDL file found
                     wdl_file = next(f for f in filenames if f.endswith('.wdl'))
+                
+                # lazy load wdl dependencies
+                import WDL
+                from cirro.models.process.wdl import get_wdl_json_schema
 
                 wdl_file = path.join(dirpath, wdl_file)
                 doc = WDL.load(wdl_file)
@@ -131,73 +134,3 @@ def get_input_params(property_path: str, definition: dict[str, Any], schema: Res
             'default': resolved.get('default', None),
             'jsonPath': jsonPath,
         }
-
-
-def get_wdl_json_schema(doc: WDL.Document) -> dict[str, Any]:  # type: ignore
-    """
-    Generate a JSON schema from a WDL Document.
-    """
-    # strictly look for top level workflow inputs
-    if not doc.workflow:
-        raise ValueError("WDL Document does not contain a workflow.")
-
-    if not doc.workflow.inputs:
-        raise ValueError("WDL Document workflow does not contain inputs.")
-
-    # only get declarations from top-level workflow inputs
-    params = {
-        child.name: {
-            'obj': child,
-            'type': str(child.type),
-            'default': child.expr.literal.value if child.expr and child.expr.literal else None,
-            'optional': child.type.optional}
-        for child in (doc.workflow.inputs or []) if isinstance(child, WDL.Decl)  # type: ignore
-    }
-
-    schema = {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "urn:cirro:wdl:schema",
-        "type": "object",
-        "properties": {},
-        "required": []
-    }
-
-    type_map = {
-        'String': 'string',
-        'Int': 'integer',
-        'Float': 'number',
-        'Boolean': 'boolean',
-        'File': 'string',  # File is typically a string path
-        'Directory': 'string',  # Directory is typically a string path
-        'Array[String]': 'array',
-        'Array[Int]': 'array',
-        'Array[Float]': 'array',
-        'Array[Boolean]': 'array',
-        'Array[File]': 'array',
-        'Array[Directory]': 'array',
-
-        # complex types
-        'Pair': 'string',  # WDL Pair type, encode as JSON string
-        'Map': 'string',  # WDL Map type, encode as JSON string
-        'Struct': 'string',  # WDL Struct type, encode as JSON string
-
-        # unsuported types
-        # 'Object': 'object',  # WDL Object type (deprecated), not directly supported in JSON Schema
-    }
-
-    # convert params into JSON schema properties
-    for name, param in params.items():
-        if param['type'].replace('?', '') not in type_map:
-            raise ValueError(f"Unsupported WDL type: {param['type']} for parameter {name}")
-
-        schema['properties'][name] = {
-            "type": type_map[param['type'].replace('?', '')],  # remove optional marker
-            "wdlType": param['type'],
-            "default": param['default'],
-            "description": f"Parameter {name} of type {param['type']}"
-        }
-        # paramters that are required have default=None and optional=False
-        if param['default'] is None and not param['optional']:
-            schema['required'].append(name)
-
-    return schema
