+++ pyproject.toml
@@ -23,7 +23,7 @@ s3fs = "2025.3.2"
 pyjwt = "^2.10.0"
 msal-extensions = "^1.0.0"
 requests-aws4auth = {version="1.3.1", optional=true}
-miniwdl = "^1.13.0"
+miniwdl = {version="^1.13.0", optional=true}
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.1.3"
@@ -33,6 +33,7 @@ pdoc = "15.0.0"
 
 [tool.poetry.extras]
 pyodide = ["requests-aws4auth"]
+wdl = ["miniwdl"]
 
 [tool.poetry.scripts]
 cirro-cli = "cirro.cli.cli:main"
